+++ spec/requests/flows/client_credentials_spec.rb
@@ -159,6 +159,47 @@ describe "Client Credentials Request" do
     end
   end
 
+  context "when revoke_previous_client_credentials_token is true" do
+    before do
+      allow(Doorkeeper.config).to receive(:reuse_access_token) { false }
+      allow(Doorkeeper.config).to receive(:revoke_previous_client_credentials_token) { true }
+    end
+
+    it "revokes the previous token" do
+      headers = authorization client.uid, client.secret
+      params  = { grant_type: "client_credentials" }
+
+      post "/oauth/token", params: params, headers: headers
+      should_have_json "access_token", Doorkeeper::AccessToken.first.token
+
+      token = Doorkeeper::AccessToken.first
+
+      post "/oauth/token", params: params, headers: headers
+      should_have_json "access_token", Doorkeeper::AccessToken.last.token
+
+      expect(token.reload.revoked?).to be_truthy
+      expect(Doorkeeper::AccessToken.last.revoked?).to be_falsey
+    end
+
+    context "with a simultaneous request" do
+      let!(:access_token) { FactoryBot.create :access_token, resource_owner_id: nil }
+
+      before do
+        allow(Doorkeeper.config.access_token_model).to receive(:matching_token_for) { access_token }
+        allow(access_token).to receive(:revoked?).and_return(true)
+      end
+
+      it "returns an error" do
+        headers = authorization client.uid, client.secret
+        params  = { grant_type: "client_credentials" }
+
+        post "/oauth/token", params: params, headers: headers
+        should_not_have_json "access_token"
+        should_have_json "error", "invalid_token_reuse"
+      end
+    end
+  end
+
   def authorization(username, password)
     credentials = ActionController::HttpAuthentication::Basic.encode_credentials username, password
     { "HTTP_AUTHORIZATION" => credentials }